Fixed onEvent parameter type.

+++ source/Actors/Actors/Punk/Punk.c
@@ -55,7 +55,7 @@ void Punk::destructor()
 
 //——————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
 
-bool Punk::onEvent(ListenerObject eventFirer __attribute__((unused)), uint32 eventCode)
+bool Punk::onEvent(ListenerObject eventFirer __attribute__((unused)), uint16 eventCode)
 {
 	switch(eventCode)
 	{
