+++ source/Components/Pong/Pong.c
@@ -72,7 +72,7 @@ typedef struct RemotePlayerData
 
 //——————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
 
-bool Pong::onEvent(ListenerObject eventFirer __attribute__((unused)), uint32 eventCode)
+bool Pong::onEvent(ListenerObject eventFirer __attribute__((unused)), uint16 eventCode)
 {
 	switch(eventCode)
 	{
