+++ source/States/CollisionsState/CollisionsState.c
@@ -29,7 +29,7 @@
 
 //——————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
 
-bool CollisionsState::onEvent(ListenerObject eventFirer __attribute__((unused)), uint32 eventCode)
+bool CollisionsState::onEvent(ListenerObject eventFirer __attribute__((unused)), uint16 eventCode)
 {
 	switch(eventCode)
 	{
