+++ source/States/PongState/PongState.c
@@ -36,7 +36,7 @@
 
 //——————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
 
-bool PongState::onEvent(ListenerObject eventFirer __attribute__((unused)), uint32 eventCode)
+bool PongState::onEvent(ListenerObject eventFirer __attribute__((unused)), uint16 eventCode)
 {
 	switch(eventCode)
 	{
