+++ source/States/SoundsState/SoundsState.c
@@ -41,7 +41,7 @@ static const SoundROMSpec* _soundSamples[] =
 
 //——————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
 
-bool SoundsState::onEvent(ListenerObject eventFirer __attribute__((unused)), uint32 eventCode)
+bool SoundsState::onEvent(ListenerObject eventFirer __attribute__((unused)), uint16 eventCode)
 {
 	switch(eventCode)
 	{
